Simplify the interval table data structure by narrowing its scope.
The simplification means that an interval table is just a set of
intervals and a set of gaps. The intervals are mapped in a one to one
way onto a set of keys. Gaps are intervals with no assignment. For our
purposes the interval table only subdivides the [0, 1] interval of
rational numbers and not API is provided to change this. There is an
interface to shrink assigned intervals producing gaps (not implemented
yet) and an interval to assign gaps to keys. This basic API will be
used to implement the fully fledged random-slicing data
structure. Separating out the interval table functionality will make
the implementation cleaner (and makes testing easier).