Implement version of rslice:owner to get multiple unique owners
Need to compute multiple owners for each key to allow for
redudancy. Repeatedly applying our random function (hash function)
until K unique owners are identified is the simplest possible way of
achieving this, and according to the random slicing paper does not
result in too much overhead/imbalance in storage. The alternative
method is to remove each owner from the slice table and repeat the
hashing process until K owners are identified. It may be possible to
achieve this without much work by walking backwards or forwards
through the interval table until K nodes are found (this is roughly
how redundancy is satisfied in consistent hashint -- walk the ring
starting at the "primary" owner and select unique nodes until K owners
are identified).